split script to run only on right page
