+++ frontend/pages/homepage/index.js
@@ -1,40 +1,186 @@
-const Http = new XMLHttpRequest();
-const url='../backend/product/GetAllProduct.php';
-Http.open("GET", url);
-Http.send();
+import {formatPrice} from '../../index.js';
 
-Http.onreadystatechange = (e) => {
-    if(Http.response){
-        let res = JSON.parse(Http.response)['data'];
-        console.log(res);
-        let content = '';
-        res.forEach(product => {
-            content += `<div class="card">
-                            <div class="inner">
-                                <img class = "card-image image" src="${product["Path"]}" alt="pc img">
-                            </div>
-                            
-                            <div class="card-body card_body">
-                                <h2 class="card-title text-center">${product['Name']}</h2>
-                                <p class="card-text card_info">i3-9100F/8GB/250GB SSD/GeForce GTX 1660 Super/Free DOS</p>
-                                
-                                <div class="card_footer">
-                                    <div class="card-stats">
-                                        <p class="card_money">${formatNumber(product['Price'])} vnd</p>
+$(document).ready(function() {
+    getTopSaleProduct();
+    getSpecialPriceProduct();
+
+    //banner owl carousel
+    $("#banner-area .owl-carousel").owlCarousel({
+        dots: true,
+        autoplay: true,
+        autoplaySpeed: 1000,
+        dotsSpeed: 1000,
+        loop: true,
+        items: 1
+    });
+
+    //new pc owl carousel
+    $("#new-pc .owl-carousel").owlCarousel({
+        loop:true,
+        nav: true,
+        dots: false,
+        responsive: {
+            0: {
+                items: 1
+            },
+            600: {
+                items: 3
+            },
+            1000: {
+                items: 5
+            }
+        }
+    })
+
+    // blogs owl carousel
+    $("#blogs .owl-carousel").owlCarousel({
+        loop: true,
+        nav: false,
+        dots: true,
+        responsive : {
+            0: {
+                items: 1
+            },
+            600: {
+                items: 3
+            }
+        }
+    })
+})
+
+function getTopSaleProduct(){
+    var request = $.get('../backend/product/GetTopSales.php',
+      function(response) {
+        if(response){
+            if(!JSON.parse(response)['error']){
+                let information = JSON.parse(response)['data'];
+                console.log("Top sale", information);
+                var list_product = information.map(function(element){
+                    return `<div class="item py-2 px-2">
+                            <div class="product font-rale">
+                                <a href="../frontend/product/${element['Id']}">
+                                    <img src="../frontend/${element['Path']}" alt="product1" class="img-fluid">
+                                </a>
+                                <div class="text-center">
+                                    <h6>
+                                        ${element['Name']}
+                                    </h6>
+                                    <div class="rating text-warning font-size-12">
+                                        <span>
+                                            <i class="fas fa-star"></i>
+                                            <i class="fas fa-star"></i>
+                                            <i class="fas fa-star"></i>
+                                            <i class="far fa-star"></i>
+                                            <i class="far fa-star"></i>
+                                        </span>
                                     </div>
-                                    <a href="#" class="btn btn-primary card_button btn-sm">More Detail</a>
+                                    <div class="price py-2">
+                                        <span>${formatPrice(element['Price'])}</span>
+                                    </div>
+                                    <button type="submit" class="btn btn-warning font-size-12" onclick="addtoCart(${element['Id']})">Add to cart</button>
                                 </div>
                             </div>
                         </div>`;
-        });
-        let container = document.getElementById('content');
-        if(container){
-            container.innerHTML = content;
+                });
+                document.getElementById('top-sale-carousel').innerHTML = list_product.join(' ');
+            }
+            else{
+                console.log("Error ", JSON.parse(response)['message']);
+            }
+            //top sale owl carousel
+            $("#top-sale .owl-carousel").owlCarousel({
+                loop:true,
+                nav: true,
+                dots: false,
+                responsive: {
+                    0: {
+                        items: 1
+                    },
+                    600: {
+                        items: 3
+                    },
+                    1000: {
+                        items: 5
+                    }
+                }
+            })
         }
-    }
+    });
+
+    // Callback handler that will be called on failure
+    request.fail(function (jqXHR, textStatus, errorThrown){
+        // Log the error to the console
+        console.error("The following error occurred: ", textStatus, errorThrown);
+    });
 }
 
+function getSpecialPriceProduct(){
+    var request = $.get('../backend/product/GetSpecialPrice.php',
+      function(response) {
+        if(response){
+            if(!response['error']){
+                let information = JSON.parse(response)['data'];
+                console.log('Special prices ', information);
+                var list_product = information.map(function(element){
+                    return `<div class="grid-item ${getProductCategory(element['CategoryId'])} border">
+                                <div class="item py-2 px-2" style="width: 200px;">
+                                    <div class="product font-rale">
+                                        <a href="../frontend/product/${element['Id']}">
+                                            <img src="../frontend/${element['Path']}" alt="product2" class="img-fluid">
+                                        </a>
+                                        <div class="text-center">
+                                            <h6>
+                                            ${element['Name']}
+                                            </h6>
+                                            <div class="rating text-warning font-size-12">
+                                                <span>
+                                                    <i class="fas fa-star"></i>
+                                                    <i class="fas fa-star"></i>
+                                                    <i class="fas fa-star"></i>
+                                                    <i class="fas fa-star"></i>
+                                                    <i class="fas fa-star-half-alt"></i>
+                                                </span>
+                                            </div>
+                                            <div class="price py-2">
+                                                <span>${formatPrice(element['Price'])}</span>
+                                            </div>
+                                            <button type="submit" class="btn btn-warning font-size-12">Add to cart</button>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>`;
+                });
+                document.getElementById('special-price-grid').innerHTML = list_product.join(' ');
+            }
+            else{
+                console.log("Error ", response['message']);
+            }
+            //isotope filter
+            var $grid = $(".grid").isotope({
+                itemSelector: '.grid-item',
+                layoutMode: 'fitRows',
+            })
+            $grid.isotope({filter: '*'})
+            //filter items on button press
+            $(".button-group").on("click", "button", function(){
+                var filterValue = $(this).attr('data-filter');
+                $grid.isotope({filter: filterValue});
+            })
+        
+        }
+    });
 
-function formatNumber(num) {
-    return num.toString().replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1.')
-  }
+    // Callback handler that will be called on failure
+    request.fail(function (jqXHR, textStatus, errorThrown){
+        // Log the error to the console
+        console.error("The following error occurred: ", textStatus, errorThrown);
+    });
+}
+
+function getProductCategory(categoryId){
+    switch(categoryId){
+        case 1: return 'Windows';
+        case 2: return 'Mac';
+        case 3: return 'Linux';
+    }
+}
