+++ frontend/pages/login/index.php
@@ -6,5 +6,6 @@ if (isset($_SESSION[USER_ID])) {
 } else {
     require_once('../frontend/components/header/header.php');
     include('../frontend/components/login/index.php');
+    echo "<script type='module' src='../frontend/pages/login/index.js'></script>";
     require_once('../frontend/components/footer/footer.php');
 }
