+++ frontend/pages/register/index.php
@@ -6,5 +6,6 @@ if (isset($_SESSION[USER_ID])) {
 } else {
     require_once('../frontend/components/header/header.php');
     include('../frontend/components/register/index.php');
+    echo "<script type='module' src='../frontend/pages/register/index.js'></script>";
     require_once('../frontend/components/footer/footer.php');
 }
