+++ frontend/pages/settings/index.php
@@ -7,5 +7,6 @@ if (!isset($_SESSION[USER_ID])) {
 } else {
     require_once('../frontend/components/header/header.php');
     include('../frontend/components/setting/index.php');
+    echo "<script type='module' src='../frontend/pages/settings/index.js'></script>";
     require_once('../frontend/components/footer/footer.php');
 }
